test: de-duplicate Ridders helpers into tests/_ridders.py
Both gradient test files carried identical copies of num_diff_ridders and
num_grad_ridders. Extract them into a shared, device-aware module
(tests/_ridders.py) and import from both. The device-aware variant (using
x.device) works on CPU and CUDA.
